Split storage pool APIs to libvirtApi/storagePool.js
